Refactor KSPAATest and KSPAAPsiTest into same test suite
After this change, you only have to add new tests to ConfigurableKSPTest,
instead of adding it to both KSPAATest and KSPAAPsiTest, which one might forget
to do.
Now they both instantiate ConfigurableKSPTest and thus always run the same
test suite with different implementations.
